+++ tests/behat/behat_local_labvirtual.php
@@ -64,7 +64,7 @@ class behat_local_labvirtual extends behat_base {
         $category = \testing_util::get_data_generator()->create_category(['name' => 'Lab Virtual category']);
 
         $batchmgr = new batch_manager();
-        $batchid  = $batchmgr->create_batch($name, (int) $teacher->id, (int) $category->id, 'Lab EAD');
+        $batchid  = $batchmgr->create_batch($name, [(int) $teacher->id], (int) $category->id, 'Lab EAD');
 
         if ((int) $count > 0) {
             $factory = new course_factory();